%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.devices.rgl-package}
\alias{R.devices.rgl-package}
\alias{R.devices.rgl}
\docType{package}

\title{Package R.devices.rgl}


\description{
  Simple functions for creating WebGL graphics (using the rgl package) and include them in dynamically generated HTML documents.
}

\section{To get started}{
  \itemize{
   \item Vignette '\href{../doc/index.html}{R.devices.rgl: Examples}'
   \item \code{\link{toWebGL}}() - evaluate graphics code and atomically save HTML WebGL code to file.
  }
}

\section{How to cite this package}{
  To cite this package, please use:
\preformatted{
@Manual{,
  title = {R.devices.rgl: Simple Functions for Creating WebGL Graphics},
  author = {Henrik Bengtsson},
  year = {2015},
  note = {R package version 0.1.1},
  url = {https://github.com/HenrikBengtsson/R.devices.rgl},
}
}
}

\author{
 Henrik Bengtsson [aut, cre, cph].
}

\section{License}{
 LGPL (>= 2.1).
}
\keyword{package}
